Reorganize JS files, rename handler
