+++ app/recordtransfer/static/recordtransfer/js/profile/htmx.js
@@ -1,40 +1,4 @@
-/**
- * Gets the current page number of the table from the active tab
- * @returns {string|null} - The current page number or null if not found
- */
-function getCurrentTablePage() {
-    // Find the active tab
-    const activeTabInput = document.querySelector(
-        ".tabs input[name=\"profile_tabs\"]:checked"
-    );
-    const activeTabContent = activeTabInput
-        ? activeTabInput.closest(".tab").nextElementSibling
-        : null;
-
-    // Find the page-info-btn within the active tab content
-    if (activeTabContent) {
-        const pageInfoBtn = activeTabContent.querySelector(".page-info-btn");
-        if (pageInfoBtn) {
-            return pageInfoBtn.getAttribute("data-current-page");
-        }
-    }
-    return null;
-}
-
-/**
- * Adds a query parameter to a URL.
- * @param {string} url - The base URL.
- * @param {string} paramName - The query parameter name.
- * @param {string|null} paramValue - The query parameter value.
- * @returns {string} - The URL with the query parameter appended if applicable.
- */
-function addQueryParam(url, paramName, paramValue) {
-    if (paramName && paramValue) {
-        const separator = url.includes("?") ? "&" : "?";
-        return `${url}${separator}${paramName}=${encodeURIComponent(paramValue)}`;
-    }
-    return url;
-}
+import { addQueryParam, getCurrentTablePage } from "./utils.js";
 
 /**
  * Handles the UI and table refresh after a delete request for an in-progress submission.
@@ -51,7 +15,7 @@ function addQueryParam(url, paramName, paramValue) {
  * @param {string} context.ID_IN_PROGRESS_SUBMISSION_TABLE - The DOM element ID of the in-progress
  * submissions table to update.
  */
-export const handleDeleteAfterRequest = (e, context) => {
+export const handleDeleteIpSubmissionAfterRequest = (e, context) => {
     // Close the modal
     const modal = document.getElementById("base_modal");
     if (modal) {
