+++ app/recordtransfer/static/recordtransfer/js/profile/utils.js
@@ -1,0 +1,38 @@
+/**
+ * Gets the current page number of the table from the active tab
+ * @returns {string|null} - The current page number or null if not found
+ */
+export function getCurrentTablePage() {
+    // Find the active tab
+    const activeTabInput = document.querySelector(
+        ".tabs input[name=\"profile_tabs\"]:checked"
+    );
+    const activeTabContent = activeTabInput
+        ? activeTabInput.closest(".tab").nextElementSibling
+        : null;
+
+    // Find the page-info-btn within the active tab content
+    if (activeTabContent) {
+        const pageInfoBtn = activeTabContent.querySelector(".page-info-btn");
+        if (pageInfoBtn) {
+            return pageInfoBtn.getAttribute("data-current-page");
+        }
+    }
+    return null;
+}
+
+/**
+ * Adds a query parameter to a URL.
+ * @param {string} url - The base URL.
+ * @param {string} paramName - The query parameter name.
+ * @param {string|null} paramValue - The query parameter value.
+ * @returns {string} - The URL with the query parameter appended if applicable.
+ */
+export function addQueryParam(url, paramName, paramValue) {
+    if (paramName && paramValue) {
+        const separator = url.includes("?") ? "&" : "?";
+        return `${url}${separator}${paramName}=${encodeURIComponent(paramValue)}`;
+    }
+    return url;
+}
+
